// src/types/appraisal.ts
interface AppraisalRequestInformation {
    companyName?: string;
    name?: string;
    shopId?: number;
    appraisalId?: number; 
    status?: string;
    type?: string;
    note?: string;
    messageForMatchingShop?: string;
    ip?: string;
    param?: string;
    customerPhone?: string;
    customerName?: string;
    customerEmail?: string;
    customerPostNumber?: string;
    customerPrefecture?: string;
    customerMunicipalities?: string;
    customerAddress?: string;
    customerAddressDetail?: string;
    carType?: string;
    carMaker?: string;
    carModelYear?: string;
    carTraveledDistance?: string;
    inspectRemain?: string;
    carState?: string;
    runnable?: string;
    wheelDrive?: string;
    fuel?: string;
    shift?: string;
    accident?: string;
    displacement?: string;
    bodyColor?: string;
    loan?: string;
    desireDate?: string;
    grade?: string;
    exterior?: string;
    scratch?: string;
    dent?: string;
    peel?: string;
    rust?: string;
    interior?: string;
    dirt?: string;
    tear?: string;
    airConditioning?: string;
    smoke?: string;
    navigation?: string;
    autoSlide?: string;
    leatherSheet?: string;
    handleType?: string;
    backMonitor?: string;
    sunroof?: string;
    wheel?: string;
    assessmentDateFirst?: string;
    assessmentDateSecond?: string;
    assessmentDateThird?: string;
    requestYmd?: string;
    createdAt?: string;
    updatedAt?: string;
    version?: number; 
    savedParameterFunction?: string;
    savedUtmParam?: string; 
    isTestData?: boolean;
    testEmailSent?: boolean;
    fpc?: string;
    photoBefore?: string;
    photoAfter?: string;
    inspectionCertPhoto?: string;
    shopTypeEnum?: string;
    shopImageUrl?: string;
    phoneNumber?: number;
    appealStatement?: string;
    address?: string;
    businessHours?: string;
}

export type {AppraisalRequestInformation };